requirements.txt
openai-whisper==20240930
flask==3.1.0
flask-swagger-ui==4.11.1
flask_cors==5.0.0
python-dotenv==1.0.1
